audio: split IHardwarweOpusDecoder, move IHardwareOpusDecoderManager
