+++ src/core/CMakeLists.txt
@@ -512,8 +512,10 @@ add_library(core STATIC
     hle/service/audio/final_output_recorder_manager_for_applet.h
     hle/service/audio/final_output_recorder_manager.cpp
     hle/service/audio/final_output_recorder_manager.h
-    hle/service/audio/hwopus.cpp
-    hle/service/audio/hwopus.h
+    hle/service/audio/hardware_opus_decoder_manager.cpp
+    hle/service/audio/hardware_opus_decoder_manager.h
+    hle/service/audio/hardware_opus_decoder.cpp
+    hle/service/audio/hardware_opus_decoder.h
     hle/service/bcat/backend/backend.cpp
     hle/service/bcat/backend/backend.h
     hle/service/bcat/bcat.cpp
